reorganized structure to avoid redundancy
