+++ rapids_cli/doctor/checks/memory.py
