+++ rapids_cli/doctor/checks/os.py
